[#10] refactor : 회원가입 약관 동의 파일 결합
체크박스와 분리해놨던 것을 다시 결합, 더 직관적으로 해석이 가능할 것 같다.
